+++ flecsi/exec/launch.hh
@@ -56,6 +56,7 @@ template<class P, class A>
 struct replace_argument<P,
   A,
   std::enable_if_t<!must_convert<std::decay_t<A>>::value>> {
+  static constexpr bool special = false;
   static A replace(A a) {
     return static_cast<A>(a);
   }
@@ -64,14 +65,17 @@ template<class P, class A>
 struct replace_argument<P,
   A,
   decltype(void(task_param<P>::replace(std::declval<A>())))> {
+  static constexpr bool special = true;
   static decltype(auto) replace(A a) {
     return task_param<P>::replace(static_cast<A>(a));
   }
 };
 
-// For each parameter-type/argument pair we have either a Color (the
-// size of a required index launch), std::monostate (for a required single
+// For each parameter-type/argument pair we have either an Index (the size of
+// a required index launch, or nothing for an empty vector), std::monostate
+// (for a required single
 // launch), or std::nullptr_t (don't care).
+using Index = std::optional<Color>;
 
 template<class P, class A>
 struct launch {
@@ -85,7 +89,7 @@ template<class P,
   class Topo,
   typename Topo::index_space S>
 struct launch<P, data::field_reference<T, L, Topo, S>> {
-  static Color get(const data::field_reference<T, L, Topo, S> & r) {
+  static Index get(const data::field_reference<T, L, Topo, S> & r) {
     return r.topology().colors();
   }
 };
@@ -95,7 +99,7 @@ template<class P,
   class Topo,
   typename Topo::index_space S>
 struct launch<P, data::multi_reference<T, L, Topo, S>> {
-  static Color get(const data::multi_reference<T, L, Topo, S> & r) {
+  static Index get(const data::multi_reference<T, L, Topo, S> & r) {
     return r.map().colors();
   }
 };
@@ -111,19 +115,25 @@ struct launch_combine {
     else {
       if constexpr(!std::is_same_v<U, std::nullptr_t>) {
         static_assert(std::is_same_v<T, U>, "implied launch types conflict");
-        if constexpr(!std::is_same_v<T, std::monostate>)
-          if(t != c.t)
+        if constexpr(!std::is_same_v<T, std::monostate>) {
+          if(!t)
+            return c;
+          if(c.t && *t != *c.t)
             flog_fatal(
-              "implied launch sizes " << t << " and " << c.t << " conflict");
+              "implied launch sizes " << *t << " and " << *c.t << " conflict");
+        }
       }
       return *this;
     }
   }
+  const T & value() const {
+    return t;
+  }
   auto get() const {
-    if constexpr(std::is_same_v<T, std::nullptr_t>)
-      return std::monostate();
+    if constexpr(std::is_same_v<T, Index>)
+      return t.value_or(0);
     else
-      return t;
+      return std::monostate();
   }
 
 private:
@@ -136,7 +146,7 @@ launch_size(std::tuple<PP...> *, const AA &... aa) {
   return (launch_combine([] {
     // An MPI task has a known launch domain:
     if constexpr(M)
-      return run::context::instance().processes();
+      return Index(run::context::instance().processes());
     else
       return nullptr;
   }()) | ... |
@@ -149,6 +159,12 @@ struct bind_base { // decomposes parameters only
 protected:
   auto visitor() {
     return [this](auto & p, auto &&) { d().visit(p); };
+  }
+
+  template<class T>
+  void visit(std::vector<T> & v) {
+    for(auto & t : v)
+      d().visit(t);
   }
 
   // The const gives a different parameter type (avoiding Clang bug #49583)
@@ -290,30 +306,52 @@ struct future<Return, exec::launch_type_t::index> {
 };
 #endif
 
-namespace exec {
+namespace exec::detail {
 template<class R>
-struct detail::task_param<future<R>> {
+struct task_param<future<R>> {
   static future<R> replace(const future<R, launch_type_t::index> &) {
     return {};
   }
 };
 template<class R>
-struct detail::must_convert<future<R, launch_type_t::index>> : std::true_type {
+struct must_convert<future<R, launch_type_t::index>> : std::true_type {};
+
+template<class P>
+struct task_param<std::vector<P>> {
+  template<class A>
+  static std::enable_if_t<replace_argument<P, const A &>::special,
+    std::vector<P>>
+  replace(const std::vector<A> & v) {
+    const util::transform_view t(v, exec::replace_argument<P, const A &>);
+    return {t.begin(), t.end()};
+  }
+};
+template<class T>
+struct must_convert<std::vector<T>> : must_convert<T> {};
+template<class P, class A>
+struct launch<std::vector<P>, std::vector<A>> {
+  using type = decltype(launch<P, A>::get(std::declval<A>()));
+  static type get(const std::vector<A> & v) {
+    launch_combine ret{type()};
+    for(auto & a : v)
+      ret = ret | launch_combine(launch<P, A>::get(a));
+    return ret.value();
+  }
 };
 
 template<class P>
-struct detail::launch<P, launch_domain> {
-  static Color get(const launch_domain & d) {
+struct launch<P, launch_domain> {
+  static Index get(const launch_domain & d) {
     return d.size_;
   }
 };
 template<class P, class T>
-struct detail::launch<P, future<T, launch_type_t::index>> {
-  static Color get(const future<T, launch_type_t::index> & f) {
+struct launch<P, future<T, launch_type_t::index>> {
+  static Index get(const future<T, launch_type_t::index> & f) {
     return f.size();
   }
 };
-} // namespace exec
+} // namespace exec::detail
 
 ///\}
 } // namespace flecsi